requirements.txt
tensorflow>=2.8.0
numpy>=1.21.0
Pillow>=8.4.0
faiss-cpu>=1.7.4
scikit-learn>=1.0.2
matplotlib>=3.5.1
opencv-python>=4.5.5
streamlit>=1.10.0
tqdm>=4.64.0
